Try to avoid racing with concurrent snap-install
We cannot use flock across virtiofsd reliably and more elaborate schemes
are not easy to implement in shell.

The actual download happens in /var/tmp and files are then copied to the
shared cache directory. Care is taken to copy just the boring bits and
not the fancy SELinux context or other xttrs that are not allowed by our
configuration of virtiofs.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>