Add support for testing CI/CD builds from Salsa
When X_SPREAD_SALSA_JOB_ID environment variable is set, the GitLab artifact for
that job is obtained, and the snapd classic package from that version is
installed. This allows running the smoke test pipeline in response to new
builds of the snapd Debian packaging.

An example job, at the time of this writing is
https://salsa.debian.org/debian/snapd/-/jobs/7311035

The job corresponds to Debian shared pipeline used by snapd packaging
repository. The "build" job builds the package from source and collects the
resulting Debian package files as artifacts.

Jira: https://warthogs.atlassian.net/browse/SNAPDENG-35065
Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>