Add support for testing CI/CD builds from Bodhi
When X_SPREAD_BODHI_ID environment variable is set, the system under test is
updated to the version of snapd contained in the advisory. This allows running
the smoke test pipeline in response to new updates proposed to Fedora and EPEL.

An example update advisory updating Fedora 42 to snapd 2.68.3 is
https://bodhi.fedoraproject.org/updates/FEDORA-2025-737127f363

Note that due to the nature of the tooling, only more recent version of snapd,
than the one available in the distribution, can be tested this way.

Jira: https://warthogs.atlassian.net/browse/SNAPDENG-35064
Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>